backend/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 智谱 AI SDK
zhipuai==2.0.1

# 向量数据库
chromadb==0.4.22
onnxruntime==1.16.3

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# 工具库
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# 数据处理
numpy==1.26.3
pandas==2.1.4

# HTTP 客户端
httpx==0.26.0
requests==2.31.0

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
